Add config flow (#25)
Adds config flow to the component, meaning it now must be configured in the Home Assistant UI under Configuration/Integrations. Benefits are UI configuration and the ability to reload the integration without restarting Home Assistant (i.e to pull in new or modified devices from HomeSeer). This functionality should be considered as an MVP and will be improved in the future - i.e. with a proper Options flow.