Update switch.py
Update is_on to use the device value.  This is needed for plugins (Example HS4 MNS Insteon) that use differnt status and control values in homeseer.  The new Insteon plugin uses 101 as on and 102 as off for control but status uses value of 0 as off and 100 as on.  Since they don't match need to look at the value if 0 then off if 1 to 100 then on.